backend/pyproject.toml
[project]
name = "llm-wiki"
version = "0.1.0"
description = "LLM Wiki Knowledge Base - Backend"
readme = {text = "", content-type = "text/plain"}
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pgvector>=0.3.0",
    # MCP Server
    "mcp[cli]>=1.20.0",
    # File storage (S3/MinIO)
    "aioboto3>=13.0.0",
    # Redis
    "redis[hiredis]>=5.0.0",
    # Document parsing
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "tree-sitter>=0.22.0",
    "markdown-it-py>=3.0.0",
    # Embedding (Ollama client)
    "httpx>=0.27.0",
    # Utilities
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
